Added convenience macros for meta-GLA info. Added UNFINISHED LogisticRegressionGLA.
Added convenience macros for defining the various meta-GLA info. The macros
define the proper information and then output themselves, so they are stable
between parsings, and the exact same macros can be used in both templated and
standard GLAs.

Added an unfinished Logistic Regression GLA that is essentially a port of
MADlib's algorithm. It does not yet have the iteration criteria, and is
probably very buggy. It is UNTESTED.