Added Generalized Filters to the system.
Pure filters are currently not supported for GFs at this time, as it will
take some additional work. Currently supported for GFs are producing a
single tuple for every input tuple, or producing zero or more tuples for
each input tuple.

The current GFWayPoint implementation borrows too much from GLASmallWayPoint,
and I feel that a lot of the logic could be condensed and brought into their
superclass, GLAWayPoint.