Completed implementation of Generalized Filters. Removed MATCH operator.
Completed enhancement of selection to support generalized filters. Two
GFs used for pattern matching were added to make up for the loss of the
MATCH operator, one using the Oniguruma library that was being used by MATCH,
and one using the C++11 Regular Expressions library.

NOTE: The system is not yet compatible with C++11. I have made some changes to
bring it closer, but it WILL NOT COMPILE under C++11 mode as of now. The reason
for this incompatibility is the use of typeof in various macros throughout the
system. typeof is a GCC extension and was replaced in C++11 by auto and
decltype. All uses of typeof need to be guarded by CPP statements.
In the future, we should check to see what C++11 features the compiler supports
and define HAS_* macros (or similar) for features we find, so that we know
what we can use.