Revamped TPC-H bulkloader. Added way of generating .pc files for sublibraries.
The TPC-H loader now uses pipes to transfer the data between dbgen and DataPath.
This requires more CPU to be used during the actual loading, but means that the
data does not need to be generated in full in flat files prior to loading. This
is useful when loading datasets that are too large to store the data twice
(once for the source files and once on the disks).

A new script has been added that will automatically make sublibraries and
create .pc files for them, and place them in a location that the system
knows about. This is for convenience.

Minor change to GroupBy to handle the case where the inner GLA returns
a state.