Modified random number generator. Added ability to use RNG functions in Piggy. Updated sampling lib.
Previously, there would be compliation errors if the RNG facilities were used
in multiple waypoints. This has now been fixed.

The parser was changed slightly to allow functions to be registered that take
no arguments. This has allowed the RNG functions to be used in Piggy queries.

The sampling library has been updated to use a new hash function (SpookyHash).